client tests passing
