+++ package.json
@@ -12,9 +12,13 @@
     "server-dev": "nodemon server/index.js",
     "lint": "eslint --ext .js --ext .jsx client/src/index.jsx server/**.js database/**.js",
     "seed": "mysql -u root -p < database/schema.sql && node ./database/seed.js",
-    "test": "jest tests/test.js --detectOpenHandles"
+    "client-test": "jest spec/client/test.js",
+    "server-test": "jest spec/server/test.js --detectOpenHandles",
+    "test": "npm run client-test && npm run server-test"
   },
   "dependencies": {
+    "@babel/plugin-syntax-jsx": "^7.10.4",
+    "@babel/plugin-transform-runtime": "^7.11.5",
     "express": "^4.17.1",
     "jquery": "^3.5.1",
     "mysql": "^2.18.1",
@@ -33,6 +37,8 @@
     "babel-loader": "^8.1.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.5",
     "eslint": "^7.9.0",
     "eslint-config-airbnb": "^18.2.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
@@ -43,6 +49,7 @@
     "http-server": "^0.12.3",
     "jest": "^26.5.2",
     "nodemon": "^2.0.4",
+    "react-test-renderer": "^16.13.1",
     "request": "^2.88.2",
     "supertest": "^5.0.0",
     "webpack": "^4.44.2",
