+++ spec/client/jest.config.js
@@ -2,9 +2,7 @@ module.exports = {
   name: 'client',
   displayName: 'Client Tests',
   verbose: true,
-  transform: { "\\.ts$": ["ts-jest"] },
-  testRegex: "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
+  testRegex: "(/__tests__/.*|\\.(test|spec))\\.(js|jsx)$",
   moduleFileExtensions: ["js", "jsx"],
-  transform: { "^.+\\.(mjs|jsx|js)$": "babel-jest" },
   transformIgnorePatterns: ['<rootDir>/node_modules/']
 };
