+++ spec/server/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  name: 'server',
+  displayName: 'Server Tests',
+  verbose: true,
+  testRegex: "(/__tests__/.*|\\.(test|spec))\\.(js|jsx)$",
+  moduleFileExtensions: ["js", "jsx"],
+  transformIgnorePatterns: ['<rootDir>/node_modules/']
+};
