+++ spec/server/server.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const path = require('path');
-const db = require('../database');
+const db = require('../../database');
 
 const app = express();
